requirements.txt
streamlit
google-generativeai
flask
python-dotenv
requests
gunicorn